ghostfidelity/requirements.txt
flask
python-dotenv
qrcode
flask_sqlalchemy
flask_migrate
